Add to cart functionality is implemented
When one tries to add the same item again, it gives warning.
When one adds an item, it gives success message.
When one adds an item, notesincart table gets updated.
Displaying items in mycart will be based on database fetching.
Displaying items functionality in mycart is ready.
Removing items fromt mycart is ready.
Frontend needs to be implemented for the rest.